refactor(signer)!: rename module `signers` to `signer` and constrain visibility
